requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
plotly
requests
textblob
vaderSentiment
wordcloud
python-dotenv
datetime
streamlit-authenticator